Added cleanup, Fix known failures, resolved scripts issue

+++ scripts/locust/locust_utils.py
@@ -85,6 +85,14 @@ class LocustUtils:
             fout.write(os.urandom(object_size))
 
     @staticmethod
+    def delete_local_obj(object_path: str):
+        if system_utils.path_exists(object_path):
+            try:
+                system_utils.remove_file(object_path)
+            except OSError as error:
+                LOGGER.error(error)
+
+    @staticmethod
     def total_time(start_time: float) -> float:
         """
         Method to calculate total time for a request to be completed
@@ -92,6 +100,11 @@ class LocustUtils:
         :return: Total time take by request
         """
         return int((time.time() - start_time) * 1000)
+
+    def get_objects(self, bucket_name):
+        bucket = self.s3_resource.Bucket(bucket_name)
+        objects = [obj.key for obj in bucket.objects.all()]
+        return objects
 
     def create_buckets(self, bucket_count: int):
         """
@@ -134,6 +147,8 @@ class LocustUtils:
                 bucket = self.s3_resource.Bucket(bucket)
                 bucket.objects.all().delete()
                 bucket.delete()
+                if bucket in self.bucket_list:
+                    self.bucket_list.pop(bucket)
                 LOGGER.info("Deleted bucket : %s", bucket)
                 events.request_success.fire(
                     request_type="delete",
@@ -157,11 +172,7 @@ class LocustUtils:
         :param bucket_name: Name of the bucket
         :param object_size: Size of the object
         """
-        if system_utils.path_exists(OBJ_NAME):
-            try:
-                system_utils.remove_file(OBJ_NAME)
-            except OSError as error:
-                LOGGER.error(error)
+        self.delete_local_obj(OBJ_NAME)
         self.create_file(object_size)
         obj_name = "test_obj{0}".format(str(time.time()))
         LOGGER.info(
@@ -175,7 +186,7 @@ class LocustUtils:
                 response_time=self.total_time(start_time),
                 response_length=10
             )
-        except (ClientError, S3UploadFailedError, FileNotFoundError)as error:
+        except (ClientError, S3UploadFailedError, FileNotFoundError) as error:
             LOGGER.error("Upload object failed with error: %s", error)
             events.request_failure.fire(
                 request_type="put",
@@ -195,27 +206,32 @@ class LocustUtils:
             bucket = self.s3_resource.Bucket(bucket_name)
             objects = [obj.key for obj in bucket.objects.all()]
             if len(objects) > 1:
-                if system_utils.path_exists(GET_OBJ_PATH):
-                    try:
-                        system_utils.remove_file(GET_OBJ_PATH)
-                    except OSError as error:
-                        LOGGER.error(error)
+                self.delete_local_obj(GET_OBJ_PATH)
                 obj_name = random.choice(objects)
                 LOGGER.info(
                     "Starting downloading the object %s form bucket %s",
                     obj_name,
                     bucket)
-                bucket.download_file(obj_name, GET_OBJ_PATH)
-                LOGGER.info(
-                    "The %s has been downloaded successfully at %s ",
-                    obj_name,
-                    GET_OBJ_PATH)
-                events.request_success.fire(
-                    request_type="get",
-                    name="download_object",
-                    response_time=self.total_time(start_time),
-                    response_length=10
-                )
+                if obj_name in self.get_objects(bucket_name):
+                    bucket.download_file(obj_name, GET_OBJ_PATH)
+                    LOGGER.info(
+                        "The %s has been downloaded successfully at %s ",
+                        obj_name,
+                        GET_OBJ_PATH)
+                    events.request_success.fire(
+                        request_type="get",
+                        name="download_object",
+                        response_time=self.total_time(start_time),
+                        response_length=10
+                    )
+                else:
+                    LOGGER.info(
+                        "The %s has been already downloaded and deleted successfully from %s ",
+                        obj_name, bucket_name)
+        except self.s3_client.exceptions.NoSuchKey:
+            LOGGER.info(
+                "Download object is not possible as key is not present on %s",
+                bucket_name)
         except ClientError as error:
             LOGGER.error("Download object failed with error: %s", error)
             events.request_failure.fire(
@@ -241,13 +257,18 @@ class LocustUtils:
                     "Deleting object %s from the bucket %s",
                     obj_name,
                     bucket)
-                self.s3_resource.Object(bucket_name, obj_name).delete()
-                events.request_success.fire(
-                    request_type="delete",
-                    name="delete_object",
-                    response_time=self.total_time(start_time),
-                    response_length=10
-                )
+                if obj_name in self.get_objects(bucket_name):
+                    self.s3_resource.Object(bucket_name, obj_name).delete()
+                    events.request_success.fire(
+                        request_type="delete",
+                        name="delete_object",
+                        response_time=self.total_time(start_time),
+                        response_length=10
+                    )
+                else:
+                    LOGGER.info(
+                        "The %s has been already deleted successfully from %s",
+                        obj_name, bucket_name)
         except ResourceNotExistsError as error:
             LOGGER.error("Delete object failed with error: %s", error)
             events.request_failure.fire(
